[spirit] Added OAM DMA into its own module
